fix(q4_hybrid): GQA head-mapping + bf16/fp16 dtype alignment for Q4_0 KV cache
Two related bugs were breaking Q4_0 decode for GQA models (e.g. Qwen3.6-35B-A3B
with 80:8 ratio, 40:4 per-TP):

1. GQA head indexing: the Q4_0 decode kernels used the query head index `h`
   directly to index KV buffer offsets. With GQA, K/V buffers only have H_KV
   heads (4 per TP), but the kernel was indexing into 40 head slots. This read
   out-of-range slots (zero-filled garbage) and the attention output was
   dominated by zeros.

   Fix: map query head `h` to the shared KV head via `kv_head = h * H_KV // H`.
   Add H_KV constexpr to both single-head and batched kernels; recompute strides
   from H_KV inside the kernel so the wrapper does not need to know about GQA.

2. bf16/fp16 dtype mismatch (root cause of garbage decode output): the Q4_0
   pool's _get_key_buffer/_get_value_buffer returned bf16 (from the Triton
   dequant kernel), while the model weights are fp16. flashinfer page table
   was configured with `kv_cache_dtype = torch.bfloat16` so it read bf16 KV
   with fp16 Q, hitting the mixed-precision paged attention path -> garbage.

   Fix 1: KV buffer getters now `.to(torch.float16)` after dequantize.
   Fix 2: `configure_kv_cache_dtype` returns `model_dtype` for q4_0 instead
   of bfloat16, so flashinfer's expected KV dtype matches the actual buffer.

Also fixes Triton causal_conv1d bf16/fp16 compile errors: prior-token loads
returned fp16 when the x_ptr is bf16, hitting a mismatch on Ampere. Add
`.to(x_ptr.dtype.element_ty)` to the `colN = tl.load(conv_states_ptrs, ...)`
calls in both KERNEL_WIDTH strategy branches.

Verified end-to-end on Qwen3.6-35B-A3B-AWQ, 2x RTX 3060 (TP=2,
--kv-cache-dtype q4_0 --attention-backend q4_hybrid --disable-cuda-graph):
model produces coherent responses (e.g. 'capital of France' -> Paris).

+++ python/sglang/srt/mem_cache/memory_pool.py
@@ -3135,7 +3135,9 @@ class MHATokenToKVPoolQ40(MHATokenToKVPool):
         from sglang.srt.layers.quantization.q40_tensor import Q40KVQuantizeUtil
 
         orig_shape = (cache_k_q40.shape[0], self.head_num, self.head_dim)
-        return Q40KVQuantizeUtil.batched_dequantize(cache_k_q40, cache_k_scale, orig_shape)
+        return Q40KVQuantizeUtil.batched_dequantize(
+            cache_k_q40, cache_k_scale, orig_shape
+        ).to(torch.float16)
 
     def _get_value_buffer(self, layer_id: int):
         cache_v_q40 = self.v_buffer[layer_id - self.start_layer]
@@ -3144,7 +3146,9 @@ class MHATokenToKVPoolQ40(MHATokenToKVPool):
         from sglang.srt.layers.quantization.q40_tensor import Q40KVQuantizeUtil
 
         orig_shape = (cache_v_q40.shape[0], self.head_num, self.v_head_dim)
-        return Q40KVQuantizeUtil.batched_dequantize(cache_v_q40, cache_v_scale, orig_shape)
+        return Q40KVQuantizeUtil.batched_dequantize(
+            cache_v_q40, cache_v_scale, orig_shape
+        ).to(torch.float16)
 
     def get_q4_kv_buffers(self, layer_id: int):
         k_packed = self.k_buffer[layer_id - self.start_layer]
